Push the span context to gracefully handle multiple routers
In the previous version, I did not anticipate the possibility that one router
might delegate to another for code-organization reasons, and so we would enter
the telemetry handler code twice while trying to service a request and lose the
parent context we attempted to save on the first time around. We witnessed this
as traces in our aggregator/visualizer that were missing root spans.

This change brings the code closer to what opentelemetry_telemetry does without
adding that dependency; namely, it maintains the parent context as a list/stack
and pushes onto it when a new span is created, popping it when a span completes.
Also, if a context already exists, we don't attempt to extract propagation from
the request headers. This change will also make the library more compatible with
webserver instrumentation like opentelemetry_cowboy, which does the header
extraction and context creation.